chore: address remaining review comments
- Fix duplicated test fixture files by consolidating them in the root tests/fixtures directory
- Fix logic inconsistency in JavaAnalyzerAgent where success was marked even with errors by using a proper warnings field
- Update documentation to use correct language identifiers and clarify the scope of format commands
- Fix test that was calling the wrong method for MVP analysis
